v1.17.1.9
7/9/17
v1.17.1.9
--------------
- Wildness of spiders and giant spiders lowered from 95% to 75% to make
easier training and obedience.
- Natural ecosystem weight lowered by 6%.
- Adjusted spawn menu to allow for quarter factors of 0.25 for higher
customizability in spider spawns.

+++ Source/Arachnophobia/ModInfo.cs
@@ -5,7 +5,6 @@ namespace Arachnophobia
     [StaticConstructorOnStartup]
     public static class ModInfo
     {
-        public static int romSpiderFactor = 1;
-
+        public static float romSpiderFactor = 1;
     }
 }
